+++ src/main/java/ru/javawebinar/topjava/repository/MealRepository.java
@@ -7,13 +7,13 @@ import java.util.List;
 
 public interface MealRepository {
 
-    Meal save(Meal meal, Long userId);
+    Meal save(Meal meal, Integer userId);
 
-    boolean delete(Long id, Long userId);
+    boolean delete(Integer id, Integer userId);
 
-    Meal get(Long id, Long userId);
+    Meal get(Integer id, Integer userId);
 
-    List<Meal> getBetween(LocalDateTime startDateTime, LocalDateTime endDateTime, Long userId);
+    List<Meal> getBetween(LocalDateTime startDateTime, LocalDateTime endDateTime, Integer userId);
 
-    List getAll(Long userId);
+    List getAll(Integer userId);
 }
